refactor: getOrCreateCategory method refactoring
